Get function start from code-start
It just might (and just did) happen that the previous globals-define
has pushed the code ptr so close to the end of the page that the next
code-start can't insert a jmp to the new page.  If we take the current
code ptr as the function start and then later call code-start, there
will be no jmp at the function start.  But if we call code-start also
to get the function start, everything will be fine.